requirements.txt
uvicorn
fastapi[standard]
requests
python-dotenv
pydantic
python-jose[cryptography] # jwt
passlib[bcrypt]           # jwt